refactor buildtable methods

+++ HabitLogger/Program.cs
@@ -51,39 +51,54 @@ void InsertHabit()
     var date = GetDateInput("Enter the date you want to log in the format dd/mm/yyyy: ");
     Console.Clear();
     
+    var habitType = GetStringInput("Please insert the habit type e.g Running, Sleeping e.t.c");
+    habitType = habitType.ToLower();
+    Console.Clear();
+    
     var quantity = GetNumberInput("Please insert habit measure of your choice in integer (no decimals allowed): ");
     Console.Clear();
-    var unit = GetUnitInput("Please insert the unit of your habit e.g litres, glasses e.t.c");
     
-    db.InsertHabit(date, quantity, unit);
+    var unit = GetStringInput("Please insert the unit of your habit e.g minutes, kilometres e.t.c");
+    unit = unit.ToLower();
+    
+    db.InsertHabit(date, quantity, unit, habitType);
+    continueMenu();
 }
 
 void GetHabit()
 {
+    if (!HasHabitsRecord())
+    {
+        continueMenu();
+        return;
+    }
+    
     GetHabits();
     var id = GetNumberInput("Enter the habit ID you wish to retrieve: ");
     var habit = db.GetHabit(id);
 
     if (habit == null)
     {
-        Console.WriteLine("No habit found!");
+        Console.WriteLine("No habit found with that ID!");
+        continueMenu();
         return;
     }
     
     Console.Clear();
     Console.WriteLine($"Habit Record");
     BuildTableHeader();
-    Console.WriteLine($"|{habit.Id,15}|{habit.Date,15}|{habit.Quantity,15}|{habit.Unit,15}|");
-    Console.WriteLine("-----------------------------------------------------------------");
+    BuildTableRows(habit);
+    continueMenu();
 }
 
 void GetHabits()
 {
     Console.Clear();
     var habits = db.GetAllHabits();
-    if (habits == null)
+    if (!habits.Any())
     {
         Console.WriteLine("No habits found!");
+        continueMenu();
         return;
     }
     
@@ -92,30 +107,60 @@ void GetHabits()
     
     foreach (var habit in habits)
     {
-        Console.WriteLine($"|{habit.Id,15}|{habit.Date,15}|{habit.Quantity,15}|{habit.Unit,15}|");
-        Console.WriteLine("-----------------------------------------------------------------");
+        BuildTableRows(habit);
     }
+
+    continueMenu();
 }
 
 void UpdateHabit()
 {
+    if (!HasHabitsRecord())
+    {
+        continueMenu();
+        return;
+    }
+    
     GetHabits();
     var id = GetNumberInput("Enter the habit ID you wish to update: ");
     Console.Clear();
     var date = GetDateInput("Enter the updated date in the format dd/mm/yyyy: ");
     Console.Clear();
     
+    var habitType = GetStringInput("Enter the updated habit type e.g Running, Sleeping e.t.c");
+    Console.Clear();
+    
     var quantity = GetNumberInput("Enter the updated habit measure of your choice in integer (no decimals allowed): ");
     Console.Clear();
-    var unit = GetUnitInput("Enter the updated unit of your habit e.g litres, glasses e.t.c");
-    db.UpdateHabit(id, date, quantity, unit);
+    var unit = GetStringInput("Enter the updated unit of your habit e.g litres, glasses e.t.c");
+    db.UpdateHabit(id, date, quantity, unit, habitType);
+    continueMenu();
 }
 
 void DeleteHabit()
 {
+    if (!HasHabitsRecord())
+    {
+        continueMenu();
+        return;
+    }
+    
     GetHabits();
     var id = GetNumberInput("Enter the habit ID you wish to delete: ");
     db.DeleteHabit(id);
+    continueMenu();
+}
+
+bool HasHabitsRecord()
+{
+    var count = db.CountHabits();
+    if (count < 1)
+    {
+        Console.WriteLine("No habits found!");
+        return false;
+    }
+
+    return true;
 }
 
 DateOnly GetDateInput(string message)
@@ -148,22 +193,34 @@ int GetNumberInput(string message)
     return quantity;
 }
 
-string GetUnitInput(string message)
+string GetStringInput(string message)
 {
     Console.WriteLine("--------------------------");
     Console.WriteLine(message);
     
-    var unit = Console.ReadLine().Trim();
+    var res = Console.ReadLine().Trim();
 
-    return unit;
+    return res;
 }
 
 void BuildTableHeader()
 {
-    Console.WriteLine("-----------------------------------------------------------------");
+    Console.WriteLine("-----------------------------------------------------------------------------------");
     Console.ForegroundColor = ConsoleColor.Green;
-    Console.WriteLine("|" + "ID".PadLeft(15) + "|" + "Date".PadLeft(15) + "|" + "Quantity".PadLeft(15) + "|" +
+    Console.WriteLine("|" + "ID".PadLeft(15) + "|" + "Date".PadLeft(15) + "|" + "Type".PadLeft(15) + "|" + "Quantity".PadLeft(15) + "|" +
                       "Unit".PadLeft(15) + "|");
     Console.ResetColor();
-    Console.WriteLine("-----------------------------------------------------------------");
+    Console.WriteLine("-----------------------------------------------------------------------------------");
+}
+
+void BuildTableRows(Habit habit)
+{
+    Console.WriteLine($"|{habit.Id,15}|{habit.Date,15}|{habit.Type,15}|{habit.Quantity,15}|{habit.Unit,15}|");
+    Console.WriteLine("-----------------------------------------------------------------------------------");
+}
+
+void continueMenu()
+{
+    Console.WriteLine("\nPress any key to continue...");
+    Console.ReadLine();
 }
